_job/requirements.txt
functions-framework
pillow==10.1.0
openai==0.28.1
requests==2.27.1
google-cloud-storage==2.13.0
pyyaml==6.0.1